readding set -e to unlocksim and handling menu close/kill correctly
This addresses stacy's comment after commit 51ffb34c (fix unlocksim)

+++ scripts/core/sxmo_dmenu.sh
@@ -17,9 +17,15 @@ case "$1" in
 	close)
 		case "$SXMO_WM" in
 			sway)
+				if ! pgrep bemenu >/dev/null; then
+					exit
+				fi
 				exec pkill bemenu >/dev/null
 				;;
 			dwm)
+				if ! pgrep dmenu >/dev/null; then
+					exit
+				fi
 				exec pkill dmenu >/dev/null
 				;;
 		esac
